+++ src/app/services/cheatsheets.service.ts
@@ -17,14 +17,21 @@ export class CheatsheetsService {
       description: 'Overview for all html entities',
       category: 'Symbols & Icons',
       icon: 'assets/icons/html5.svg',
-      url: 'html-entities',
+      url: '/cheatsheets/html-entities'
+    },
+    {
+      name: 'UTF-8 Symbols',
+      description: 'Overview for all utf-8 symbols',
+      category: 'Symbols & Icons',
+      icon: 'assets/icons/html5.svg',
+      url: '/cheatsheets/utf-eight-symbols'
     },
     {
       name: 'CSS Cursors',
       description: 'Overview of all available css cursors',
       category: 'Symbols & Icons',
-      icon: 'assets/icons/html5.svg',
-      url: 'css-cursor',
+      icon: 'assets/icons/cursor.svg',
+      url: '/cheatsheets/css-cursors'
     }
   ];
 }