+++ src/app/tools/base64/base64.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FavoritesService } from 'src/app/services/favorites.service';
+import { Clipboard } from '@angular/cdk/clipboard';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-base64',
@@ -10,12 +12,25 @@ export class Base64Component implements OnInit {
   input: string = '';
   output: string = '';
   action: string = 'encode';
+  type: string = 'txt';
+  selectedImg: any;
 
   showOutput = false;
   isFav = false;
+  item = {
+    name: 'Base64',
+    description: 'Encode and decode strings or images with Base64',
+    category: 'Encoding/Decoding',
+    icon: 'assets/icons/base64.svg',
+    url: '/tools/base64',
+  };
 
-  constructor(private favS: FavoritesService) {
-    this.isFav = this.favS.isFavorite('base64');
+  constructor(
+    private favS: FavoritesService,
+    private clipboard: Clipboard,
+    private _snackBar: MatSnackBar
+  ) {
+    this.isFav = Boolean(this.favS.isFavorite(this.item));
   }
 
   ngOnInit(): void {}
@@ -29,24 +44,33 @@ export class Base64Component implements OnInit {
   }
 
   addToFav() {
-    let fav = {
-      name: 'Base64',
-      description: 'Encode and decode strings or images with Base64',
-      category: 'Encoding/Decoding',
-      icon: 'assets/icons/base64.svg',
-      url: 'base64',
-    };
-
     this.isFav = true;
-    this.favS.addToFavorites(fav);
+    this.favS.addToFavorites(this.item);
   }
 
   handleInput() {
     this.showOutput = true;
-    if (this.action.toLowerCase() == 'decode') {
-      this.output = this.fromBase64(this.input);
-    } else {
-      this.output = this.toBase64(this.input);
+
+    if (this.type == 'txt') {
+      if (this.action.toLowerCase() == 'decode') {
+        this.output = this.fromBase64(this.input);
+      } else {
+        this.output = this.toBase64(this.input);
+      }
     }
+  }
+
+  setImage(event: any) {
+    const file = event.target.files[0];
+    const reader = new FileReader();
+    reader.readAsDataURL(file);
+    reader.onload = () => {
+      this.output = reader.result + '';
+    };
+  }
+
+  copyText() {
+    this.clipboard.copy(this.output);
+    this._snackBar.open('Successfully copied.', 'Close');
   }
 }