+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.0]
+
+### Added
+ - Landing/Home Page
+ - Base64 encoding for images
+ - Device Info tool xxxxx may mehr?
+ - Text To Json Converter xxxxxx not done
+ - CSS Cursors Cheatsheet
+ - UTF-8 Symbols Cheatsheet xxxxxx not done
+ - Tailwindcss Libary
+
+### Changed
+ - Fixed error with function to save tool/cheatsheet as favorite (prev. only one was saved due to bad code 🤷🏼‍♂️)
+
 ## [0.0.2]
 
 ### Added
@@ -11,7 +25,6 @@ All notable changes to this project will be documented in this file.
  - Cheatsheets Overview Page
  - Cheatsheet Interface and Service
  - HTML Entitiew Cheatsheet
- - CSS Cursor Cheatsheet
 
  ### Changed
  - Replaced Angular Material mat-toolbar with bootstrap nav